feat: modularize brand token kits
